Fix model/schema mismatches found by auditing against core-openapi.json
- Fix 10 *List models modeled as `class X(BaseModel): RootModel: Y` (same bug
  as ClusterReallocation) to `RootModel[list[Y]]`: IamUserPolicyList,
  IamUserList, ProcessList, FilesystemFileList, FilesystemList, RtmpList,
  ClusterNodeList, ClusterDbLockList, ConfigStorageS3List, ReportProcessList
- Fix Srt.socketid from str to int (OpenAPI api.SRTChannel)
- Fix ProcessStateProgressIOTee.fifo_recovery_attempts_total from int to float
  (OpenAPI api.ProgressIOTee); fractional values no longer raise

A field-type audit of all models against the OpenAPI schemas now reports zero
remaining mismatches. Add regression unit tests.

+++ core_client/base/models/v3/process_state_progress_io_tee.py
@@ -22,7 +22,7 @@ class ProcessStateProgressIOTee(BaseModel):
     state: str | None = None  # e.g., "running", "stopped", "error"
     fifo_enabled: bool = False
     fifo_state: str | None = None  # e.g., "running", "stopped", "error"
-    fifo_recovery_attempts_total: int | None = None
+    fifo_recovery_attempts_total: float | None = None
 
     class ConfigDict:
         use_enum_values = True
